Service - Clean up client connections cde
